Basic version functioning fully

+++ gpu_expInt.h
@@ -8,7 +8,7 @@ extern void GPUexponentialIntegralFloat(float *results, int block_size_X, int bl
 
 #define _GPU_EXP_INT_H_
 
-__device__ float calcExp_slow(int n, int x, int maxIters);
+__device__ float calcExp_slow(int n, float x, int maxIters);
 
 __global__ void calcExpIntegralGrid(float *res_glob, int n, int numSamples, int a, float division, int maxIters);
